refactor: Rename serialization package to json
