+++ src/test/java/de/rub/nds/modifiablevariable/json/ByteSerializationTest.java
@@ -5,7 +5,7 @@
  *
  * Licensed under Apache License 2.0 http://www.apache.org/licenses/LICENSE-2.0
  */
-package de.rub.nds.modifiablevariable.serialization;
+package de.rub.nds.modifiablevariable.json;
 
 import static org.junit.jupiter.api.Assertions.*;
